backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
boto3>=1.34
pyjwt[crypto]>=2.8
